speed up weighted sum for model_builder and sat python; more work on lrat; speed improvement on probing; add TrivialLiteral repository

+++ ortools/sat/presolve_util.cc
@@ -591,8 +591,8 @@ int ActivityBoundHelper::RemoveEnforcementThatMakesConstraintTrivial(
                           amo_indices_[negated_index].end());
     ++num_enforcement_to_check;
   }
-  int non_amo_min_activity = 0;
-  int non_amo_max_activity = 0;
+  int64_t non_amo_min_activity = 0;
+  int64_t non_amo_max_activity = 0;
   auto log_work = [&]() {
     VLOG(1) << "RemoveEnforcementThatMakesConstraintTrivial: "
                "aborting because too expensive: "
